+++ services/rds/v3/model/model_postgresql_list_database.go
@@ -22,7 +22,7 @@ type PostgresqlListDatabase struct {
 	CollateSet *string `json:"collate_set,omitempty"`
 	// 数据库大小（单位：字节）。
 
-	Size *int32 `json:"size,omitempty"`
+	Size *int64 `json:"size,omitempty"`
 }
 
 func (o PostgresqlListDatabase) String() string {
